requirements.txt
Appium-Python-Client==2.2.0
async-generator==1.10
attrs==21.4.0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.2
docopt==0.6.2
google-auth==2.6.5
google-auth-oauthlib==0.5.1
google-oauth==1.0.1
gspread==5.3.2
gspread-dataframe==3.3.0
h11==0.13.0
idna==3.3
numpy==1.22.3
oauthlib==3.2.0
outcome==1.1.0
pandas==1.4.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
selenium==4.1.3
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
urllib3==1.26.9
wsproto==1.1.0
